Refactor into dropbox module
