+++ .idea/gradle.xml
@@ -11,6 +11,7 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/app/auth" />
+            <option value="$PROJECT_DIR$/app/dropbox" />
             <option value="$PROJECT_DIR$/app/filemanager" />
           </set>
         </option>
