+++ app/dropbox/src/main/java/com/duchastel/simon/photocategorizer/dropbox/di/DropboxQualifier.kt
@@ -1,0 +1,7 @@
+package com.duchastel.simon.photocategorizer.dropbox.di
+
+import javax.inject.Qualifier
+
+@Qualifier
+@Retention(AnnotationRetention.BINARY)
+annotation class Dropbox
