+++ app/dropbox/src/test/java/com/duchastel/simon/photocategorizer/dropbox/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package com.duchastel.simon.photocategorizer.dropbox
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
