+++ settings.gradle.kts
@@ -23,3 +23,4 @@ rootProject.name = "Photo Categorizer"
 include(":app")
 include(":app:auth")
 include(":app:filemanager")
+include(":app:dropbox")
